backend/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
requests==2.32.5
yt-dlp==2025.10.22
supabase==2.22.2
python-dotenv==1.0.1